WIP [MSTSC] Fix warnings

+++ base/applications/mstsc/iso.c
@@ -60,7 +60,7 @@ static void
 iso_send_connection_request(char *username, uint32 neg_proto)
 {
 	STREAM s;
-	int length = 30 + strlen(username);
+	intptr_t length = 30 + strlen(username);
 
 	if (g_rdp_version >= RDP_V5 && g_negotiate_rdp_protocol)
 		length += 8;
